requirements.txt
langchain-google.genai
langchain
langchain_experimental
pandas
tabulate
wikipedia
langgraph
tavily-python
pydantic_ai
logfire
langchain_community 
tiktoken  
langchainhub 
chromadb 
langchain 
langgraph  
tavily-python
python-dotenv
google-generativeai
langchain-google-genai
langchain-groq
langchain-nomic
langchain-text-splitters
langchain_mistralai
wikipedia
langchain_huggingface
google-search-results
faiss-cpu
sentence_transformers
duckduckgo-search
yfinance